requirements.txt
APScheduler==3.11.1
asgiref==3.10.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
ConfigArgParse==1.7.1
contourpy==1.3.3
cycler==0.12.1
Django==5.2.8
djangorestframework==3.16.1
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
fonttools==4.60.1
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
locust==2.39.1
locust-cloud==1.26.3
MarkupSafe==3.0.2
matplotlib==3.10.7
msgpack==1.1.1
narwhals==2.11.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==12.0.0
platformdirs==4.4.0
plotly==6.4.0
psutil==7.0.0
pycparser==2.22
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pywin32==311
pyzmq==27.0.2
requests==2.32.5
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.1.1
zope.interface==7.2
